Add files via upload
In this web scraping with Python code, the data (basically the news headlines) are collected from Google news, displayed on the console window as well as a copy of the same will be sent to your e-mail at the same time. On every execution, this Python code scrapes the news headlines and do some custom formatting before saving it to notepad (or text file). In the consequent stage, the e-mail is composed by accessing the text file and finally sent to the mentioned e-mail address.